requirements.txt
## NoneBot2
nonebot2
nonebot-adapter-onebot

aiohttp
psutil
Wappalyzer
python-Wappalyzer
python-whois
playwright
nonebot_plugin_apscheduler
distro
setuptools
wmi
gputil

# selenium
# webdriver-manager
# pyppeteer-stealth
#pycuda